services/workshop/requirements.txt
Django~=2.2.4
django-environ==0.4.5
psycopg2-binary==2.8.5
PyJWT==1.7.1
requests==2.24.0
django-cors-headers==3.4.0
bcrypt==3.1.7
djangorestframework==3.11.0
django-health-check==3.12.1
django-extended-choices==1.3.3
djongo==1.3.3